# frontend/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/api/auth/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterResponseDto"}}}},"400":{"description":"Bad request"}},"summary":"Register a new user","tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"User logged in successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}},"400":{"description":"Bad request"},"401":{"description":"Unauthorized"}},"summary":"Login a user","tags":["Auth"]}},"/api/auth/profile":{"get":{"operationId":"AuthController_getProfile","parameters":[],"responses":{"200":{"description":"User profile retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProfileResponseDto"}}}},"401":{"description":"Unauthorized"}},"summary":"Get user profile","tags":["Auth"]}},"/api/tenants":{"post":{"operationId":"TenantController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTenantDto"}}}},"responses":{"201":{"description":"The tenant has been successfully created.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"400":{"description":"Bad request."},"409":{"description":"Tenant with this name already exists."}},"security":[{"bearer":[]}],"summary":"Create a new tenant","tags":["tenants"]},"get":{"operationId":"TenantController_findAll","parameters":[],"responses":{"200":{"description":"Return all tenants.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Tenant"}}}}}},"security":[{"bearer":[]}],"summary":"Get all tenants","tags":["tenants"]}},"/api/tenants/{id}":{"get":{"operationId":"TenantController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the tenant.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"404":{"description":"Tenant not found."}},"security":[{"bearer":[]}],"summary":"Get a tenant by ID","tags":["tenants"]},"patch":{"operationId":"TenantController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTenantDto"}}}},"responses":{"200":{"description":"The tenant has been successfully updated.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"404":{"description":"Tenant not found."},"409":{"description":"Tenant with this name already exists."}},"security":[{"bearer":[]}],"summary":"Update a tenant","tags":["tenants"]},"delete":{"operationId":"TenantController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The tenant has been successfully deleted."},"404":{"description":"Tenant not found."}},"security":[{"bearer":[]}],"summary":"Delete a tenant","tags":["tenants"]}},"/api/tenants/by-slug/{slug}":{"get":{"operationId":"TenantController_getTenantBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Return the tenant.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Tenant"}}}},"404":{"description":"Tenant not found."}},"security":[{"bearer":[]}],"summary":"Get a tenant by slug","tags":["tenants"]}},"/api/inventory":{"post":{"operationId":"InventoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateItemDto"}}}},"responses":{"201":{"description":"Item created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Create a new inventory item","tags":["inventory"]},"get":{"operationId":"InventoryController_findAll","parameters":[],"responses":{"200":{"description":"Items retrieved successfully","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/InventoryResponseDto"}}}}}},"security":[{"bearer":[]}],"summary":"Get all inventory items for the tenant","tags":["inventory"]}},"/api/inventory/{id}":{"get":{"operationId":"InventoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Inventory item ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Item retrieved successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Get a specific inventory item","tags":["inventory"]},"patch":{"operationId":"InventoryController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Inventory item ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateItemDto"}}}},"responses":{"200":{"description":"Item updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InventoryResponseDto"}}}}},"security":[{"bearer":[]}],"summary":"Update an inventory item","tags":["inventory"]},"delete":{"operationId":"InventoryController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Inventory item ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Item deleted successfully"}},"security":[{"bearer":[]}],"summary":"Delete an inventory item","tags":["inventory"]}},"/api/inventory/stock/check":{"get":{"operationId":"InventoryController_checkStockLevels","parameters":[],"responses":{"200":{"description":"Stock levels checked successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockCheckDto"}}}}},"security":[{"bearer":[]}],"summary":"Check for low stock items","tags":["inventory"]}},"/api/categories":{"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":"Category successfully created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CategoryResponseDto"}}}},"400":{"description":"Bad Request"},"403":{"description":"Forbidden"}},"security":[{"bearer":[]}],"summary":"Create a new category","tags":["categories"]},"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":"Returns list of categories","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDtoWithoutCreatedBy"}}}}}},"security":[{"bearer":[]}],"summary":"Get all categories for the current tenant","tags":["categories"]}},"/api/categories/tree":{"get":{"operationId":"CategoryController_getCategoryTree","parameters":[],"responses":{"200":{"description":"Returns category tree","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDtoWithoutCreatedBy"}}}}}},"security":[{"bearer":[]}],"summary":"Get category tree for the current tenant","tags":["categories"]}},"/api/categories/{id}":{"get":{"operationId":"CategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","description":"Category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the category","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryResponseDto"}}}},"404":{"description":"Category not found"}},"security":[{"bearer":[]}],"summary":"Get a specific category by ID","tags":["categories"]},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","description":"Category ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":"Category updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryResponseDto"}}}},"400":{"description":"Bad Request"},"403":{"description":"Forbidden"},"404":{"description":"Category not found"}},"security":[{"bearer":[]}],"summary":"Update a category","tags":["categories"]},"delete":{"operationId":"CategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","description":"Category ID","schema":{"type":"string"}}],"responses":{"200":{"description":"Category deleted successfully"},"400":{"description":"Cannot delete category with items or children"},"403":{"description":"Forbidden"},"404":{"description":"Category not found"}},"security":[{"bearer":[]}],"summary":"Delete a category","tags":["categories"]}}},"info":{"title":"Inventory API","description":"The inventory management API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpDto":{"type":"object","properties":{"username":{"type":"string"},"role":{"type":"string"},"tenantName":{"type":"string"}},"required":["username","role","tenantName"]},"TenantResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"slug":{"type":"string"}},"required":["id","name","slug"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"role":{"type":"string"},"tenant":{"$ref":"#/components/schemas/TenantResponseDto"}},"required":["id","username","role","tenant"]},"RegisterResponseDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserResponseDto"},"access_token":{"type":"string"}},"required":["user","access_token"]},"LoginDto":{"type":"object","properties":{"username":{"type":"string"}},"required":["username"]},"LoginResponseDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserResponseDto"},"access_token":{"type":"string"}},"required":["user","access_token"]},"ProfileResponseDto":{"type":"object","properties":{"data":{"$ref":"#/components/schemas/UserResponseDto"}},"required":["data"]},"CreateTenantDto":{"type":"object","properties":{}},"Tenant":{"type":"object","properties":{}},"UpdateTenantDto":{"type":"object","properties":{}},"CreateItemDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"},"sku":{"type":"string"},"minStockLevel":{"type":"number"},"maxStockLevel":{"type":"number"},"categoryId":{"type":"string"}},"required":["name","description","quantity","price","sku","minStockLevel","maxStockLevel","categoryId"]},"InventoryResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"},"sku":{"type":"string"},"minStockLevel":{"type":"number"},"categoryId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","name","description","quantity","price","sku","minStockLevel","categoryId","createdAt","updatedAt"]},"UpdateItemDto":{"type":"object","properties":{"name":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"},"sku":{"type":"string"},"minStockLevel":{"type":"number"},"maxStockLevel":{"type":"number"},"categoryId":{"type":"string"}},"required":["name","quantity","price","sku","minStockLevel","maxStockLevel","categoryId"]},"StockCheckDto":{"type":"object","properties":{"lowStock":{"type":"array","items":{"$ref":"#/components/schemas/InventoryResponseDto"}}},"required":["lowStock"]},"CreateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"parentId":{"type":"string"}},"required":["name","description","parentId"]},"UserResponseWithTenantIdDto":{"type":"object","properties":{"id":{"type":"string"},"username":{"type":"string"},"role":{"type":"string"},"tenantId":{"type":"string"}},"required":["id","username","role","tenantId"]},"CategoryResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tenantId":{"type":"string"},"parent":{"$ref":"#/components/schemas/CategoryResponseDto"},"parentId":{"type":"string"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdById":{"type":"string"},"createdBy":{"$ref":"#/components/schemas/UserResponseWithTenantIdDto"}},"required":["id","name","description","tenantId","parent","parentId","children","createdAt","updatedAt","createdById","createdBy"]},"CategoryResponseDtoWithoutCreatedBy":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tenantId":{"type":"string"},"parent":{"$ref":"#/components/schemas/CategoryResponseDto"},"parentId":{"type":"string"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdById":{"type":"string"}},"required":["id","name","description","tenantId","parent","parentId","children","createdAt","updatedAt","createdById"]},"InventoryItemDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tenantId":{"type":"string"},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"minStockLevel":{"type":"number"},"categoryId":{"type":"string"},"category":{"$ref":"#/components/schemas/CategoryResponseDto"}},"required":["id","name","description","tenantId","createdAt","updatedAt","minStockLevel","categoryId","category"]},"UpdateCategoryResponseDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tenantId":{"type":"string"},"parent":{"$ref":"#/components/schemas/CategoryResponseDto"},"parentId":{"type":"string"},"children":{"type":"array","items":{"$ref":"#/components/schemas/CategoryResponseDto"}},"createdAt":{"format":"date-time","type":"string"},"updatedAt":{"format":"date-time","type":"string"},"createdById":{"type":"string"},"createdBy":{"$ref":"#/components/schemas/UserResponseWithTenantIdDto"},"items":{"type":"array","items":{"$ref":"#/components/schemas/InventoryItemDto"}}},"required":["id","name","description","tenantId","parent","parentId","children","createdAt","updatedAt","createdById","createdBy"]},"UpdateCategoryDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"parentId":{"type":"string"}}}}}}